refactor(consensus): remove `inner` module
Simplify and flatten the module structure in `consensus` by inlining the
`inner` submodule.

The `inner` submodule had been used to feature gate the consensus task
implementation. There were two different versions of it, behind a
feature flag check. Since
https://github.com/equilibriumco/pathfinder/pull/3344, moved the feature
gating outside the consensus module, the `inner` submodule does nothing
but increase the nesting of the module structure. Thus, remove it.
